Add content-type check
